Rename RcFileCorruptionException and move to shared package
