+++ lib/trino-hive-formats/src/main/java/io/trino/hive/formats/rcfile/text/ListEncoding.java
@@ -15,7 +15,7 @@ package io.trino.hive.formats.rcfile.text;
 
 import io.airlift.slice.Slice;
 import io.airlift.slice.SliceOutput;
-import io.trino.hive.formats.rcfile.RcFileCorruptionException;
+import io.trino.hive.formats.FileCorruptionException;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.type.Type;
@@ -33,7 +33,7 @@ public class ListEncoding
 
     @Override
     public void encodeValueInto(int depth, Block block, int position, SliceOutput output)
-            throws RcFileCorruptionException
+            throws FileCorruptionException
     {
         byte separator = getSeparator(depth);
 
@@ -53,7 +53,7 @@ public class ListEncoding
 
     @Override
     public void decodeValueInto(int depth, BlockBuilder builder, Slice slice, int offset, int length)
-            throws RcFileCorruptionException
+            throws FileCorruptionException
     {
         byte separator = getSeparator(depth);
         int end = offset + length;
@@ -79,7 +79,7 @@ public class ListEncoding
     }
 
     private void decodeElementValueInto(int depth, BlockBuilder blockBuilder, Slice slice, int offset, int length)
-            throws RcFileCorruptionException
+            throws FileCorruptionException
     {
         if (nullSequence.equals(0, nullSequence.length(), slice, offset, length)) {
             blockBuilder.appendNull();
